Listen for prequal:close event to close modal
The Javascript models that are used to display the educational modals
have recently changed. This broke the integration with the modal and
Android SDK since the modal change was not backwards-compatible. An
event ("prequal:close") is emitted when the new modal is closed, and
this commit listens for that event and calls the modal close callback.